Update README.md
A remote controlled vehicle is any machine controlled by means that is physically not connected with origin external to the machine. There are many types in it, based on the controls – radio control device, Wi-Fi controlled and even Bluetooth controlled. These devices are always controlled by umans and take no action autonomously. The main target in such vehicles would be to safely reach a designated point, the area and reach back to the point of origin